move code to settings folder
