+++ src/components/Providers/ClientProviders.tsx
@@ -5,7 +5,6 @@ import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
 import { darkTheme, lightTheme } from 'lib/MuiTheme';
 import { useTheme as useNextTheme } from 'next-themes';
 import { SnackbarProvider } from 'notistack';
-import { useLayoutEffect, useState } from 'react';
 
 type ClientProvidersProps = {
   children: ReactNode;
@@ -13,18 +12,7 @@ type ClientProvidersProps = {
 
 const ClientProviders = ({ children }: ClientProvidersProps) => {
   const { theme: nextTheme } = useNextTheme();
-  const [mounted, setMounted] = useState(false);
-
-  useLayoutEffect(() => {
-    setMounted(true);
-  }, []);
-
-  if (!mounted) {
-    return <>{children}</>;
-  }
-
-  const muiTheme
-    = nextTheme === 'dark' || nextTheme === 'system' ? darkTheme : lightTheme;
+  const muiTheme = nextTheme === 'dark' ? darkTheme : lightTheme;
 
   return (
     <MuiThemeProvider theme={muiTheme}>
