+++ src/messages/en.json
@@ -18,7 +18,7 @@
   },
   "PlannerPage": {
     "year-with-value": "Year {value}",
-    "add-a-year": "Add a year",
+    "add-year": "Add Year",
     "sessionTerms": {
       "AUTOMNE": "Autumn",
       "HIVER": "Winter",
